PW-200: Refactor test file structure
