+++ tests/saucedemo.spec.ts
@@ -1,4 +1,3 @@
-
 import { test, expect } from './base';
 
 test.beforeEach(async ({ page }) => {
